[CRK-3] integrate with pagination for load more and mapper data into pagedList

+++ shared/src/main/java/com/patipan/dev/shared/service/CoinRankingService.kt
@@ -10,6 +10,6 @@ interface CoinRankingService {
     @GET("v1/public/coins")
     fun getListCoinRanking(
         @Query("prefix") searchPrefix: String? = null,
-        @Query("offset") page: Int? = null
+        @Query("offset") page: Long? = null
     ): Call<CoinRankingResponse>
 }
